requirements.txt

transformers
torch
sentence-transformers
langchain
faiss-cpu
chromadb
fastapi
uvicorn
aif360
shap
xgboost
openai
pillow
gradio
scipy
scikit-learn
datasets
